# openapi.yaml
openapi: 3.0.0
info:
  title: Fit & Fun API
  description: API de suivi de performances sportives et de données biométriques
  version: 1.0.0
...
paths:
...
  /activities:
    post:
      tags:
        - Activity
      summary: Enregistrer une séance
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Activity'
      responses:
        '201':
          description: Activité créée
        '400':
          description: JSON invalide
        '404':
          description: User inconnu (via user_id)
        '422':
          description: Logique métier
  /activities/{id}:
    get:
      tags:
        - Activity
      summary: Détails d'une séance
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Succès
        '404':
          description: Activité inexistante
    delete:
      tags:
        - Activity
      summary: Supprimer une séance
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Supprimé
        '404':
          description: Activité inexistante
...
components:
  schemas:
...
    Activity:
      type: object
      required: [user_id, type, duration, calories, timestamp]
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        user_id:
          type: string
          format: uuid
        type:
          type: string
          description: 'Exemple: foot, basket, salle...'
        duration:
          type: integer
        calories:
          type: integer
        timestamp:
          type: string
          format: date-time